\name{Exact one-sided and two-sided 1-alpha confidence interval, main functions}
\alias{BinomialCIone}
\alias{PairedCIone}
\title{Inductive confidence intervals for the difference between two proportions}
\description{
The function BinomialCIone and PairedCIone are the called functions for the main function BinomCI and PairedCI, respectively.         
}